[all] add type support
Modules can now specify which types of events they support. This
also means that users can specify which types of events they will
require and that libverto will return only an implementation that
supports the required features. This also adds a function which
will permit you to query the features supported after a context
is already created.